+++ src/superannotate/lib/core/entities/__init__.py
@@ -4,6 +4,7 @@ from lib.core.entities.base import SubSetEntity
 from lib.core.entities.classes import AnnotationClassEntity
 from lib.core.entities.folder import FolderEntity
 from lib.core.entities.integrations import IntegrationEntity
+from lib.core.entities.items import CategoryEntity
 from lib.core.entities.items import ClassificationEntity
 from lib.core.entities.items import DocumentEntity
 from lib.core.entities.items import ImageEntity
@@ -12,7 +13,6 @@ from lib.core.entities.items import PROJECT_ITEM_ENTITY_MAP
 from lib.core.entities.items import TiledEntity
 from lib.core.entities.items import VideoEntity
 from lib.core.entities.project import AttachmentEntity
-from lib.core.entities.project import CategoryEntity
 from lib.core.entities.project import ContributorEntity
 from lib.core.entities.project import CustomFieldEntity
 from lib.core.entities.project import ProjectEntity